LayoutTests/fast/backgrounds/background-position-parsing-2-expected.txt

Test to make sure background-position is parsed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS style.backgroundPosition is '70% 50%'
PASS computedStyle.backgroundPosition is '70% 50%'
PASS style.backgroundPosition is '84px 50%'
PASS computedStyle.backgroundPosition is '84px 50%'
PASS style.backgroundPosition is '0% 50%'
PASS computedStyle.backgroundPosition is '0% 50%'
PASS style.backgroundPosition is '100% 50%'
PASS computedStyle.backgroundPosition is '100% 50%'
PASS style.backgroundPosition is '50% 100%'
PASS computedStyle.backgroundPosition is '50% 100%'
PASS style.backgroundPosition is '50% 0%'
PASS computedStyle.backgroundPosition is '50% 0%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '0% 100%'
PASS computedStyle.backgroundPosition is '0% 100%'
PASS style.backgroundPosition is '0% 100%'
PASS computedStyle.backgroundPosition is '0% 100%'
PASS style.backgroundPosition is '100% 100%'
PASS computedStyle.backgroundPosition is '100% 100%'
PASS style.backgroundPosition is '100% 0%'
PASS computedStyle.backgroundPosition is '100% 0%'
PASS style.backgroundPosition is '54px 100%'
PASS computedStyle.backgroundPosition is '54px 100%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '5% 100%'
PASS computedStyle.backgroundPosition is '5% 100%'
PASS style.backgroundPosition is '30pt -20px'
PASS computedStyle.backgroundPosition is '40px -20px'
PASS style.backgroundPosition is '100% 50%'
PASS computedStyle.backgroundPosition is '100% 50%'
PASS style.backgroundPosition is '100% 0px'
PASS computedStyle.backgroundPosition is '100% 0px'
PASS style.backgroundPosition is '100% 50%'
PASS computedStyle.backgroundPosition is '100% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '0% 50%'
PASS computedStyle.backgroundPosition is '0% 50%'
PASS style.backgroundPosition is '-20% 50%'
PASS computedStyle.backgroundPosition is '-20% 50%'
PASS style.backgroundPosition is '100% 0%'
PASS computedStyle.backgroundPosition is '100% 0%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% 50%, 100% 50%'
PASS computedStyle.backgroundPosition is '50% 50%, 100% 50%'
PASS style.backgroundPosition is '50% 0%, 50% 100%'
PASS computedStyle.backgroundPosition is '50% 0%, 50% 100%'
PASS style.backgroundPosition is '100% 50%, 5px 100%'
PASS computedStyle.backgroundPosition is '100% 50%, 5px 100%'
PASS style.backgroundPosition is '0% 0%, 100% 100%'
PASS computedStyle.backgroundPosition is '0% 0%, 100% 100%'
PASS style.backgroundPosition is '0% 0%, 100% 100%'
PASS computedStyle.backgroundPosition is '0% 0%, 100% 100%'
PASS style.backgroundPosition is '0% 0%, 100% 100%'
PASS computedStyle.backgroundPosition is '0% 0%, 100% 100%'
PASS style.backgroundPosition is '0% 0%, 100% 100%'
PASS computedStyle.backgroundPosition is '0% 0%, 100% 100%'
PASS successfullyParsed is true

TEST COMPLETE
